+++ src/lib/core/LightboxClient/assemblers.ts
@@ -1,4 +1,5 @@
-import { Styles, Ids } from '@/constants';
+import { Config } from '@/config';
+import { Styles, Ids, Dimensions } from '@/constants';
 import { ClientCallbacks, ApiStructure, ClientStyles } from '@/types';
 
 export const mountListener = (callbacks: ClientCallbacks) => {
@@ -34,12 +35,18 @@ const mountStyles = (styles: ClientStyles) => {
         ?.map((x) => parseInt(x, 16)) ?? [107, 114, 128];
 
     background.push(styles.background?.opacity ?? 0.75);
-
     document.documentElement.style.setProperty(Styles.BACKGROUND_COLOR, `rgb(${background.join(', ')})`);
 
     const rounded = styles.rounded ?? 0;
-
     document.documentElement.style.setProperty(Styles.ROUNDED, `${rounded.toString()}px`);
+
+    let height = styles.height ?? Config.defaultDimension.height;
+    if (styles.dimension) height = Dimensions[styles.dimension.toUpperCase() as keyof typeof Dimensions].height;
+    document.documentElement.style.setProperty(Styles.MAX_HEIGHT, `${height.toString()}px`);
+
+    let width = styles.width ?? Config.defaultDimension.width;
+    if (styles.dimension) width = Dimensions[styles.dimension.toUpperCase() as keyof typeof Dimensions].width;
+    document.documentElement.style.setProperty(Styles.MAX_WIDTH, `${width.toString()}px`);
 };
 
 const unmountLightbox = (listener: (event: MessageEvent<ApiStructure>) => void) => {