+++ src/lib/core/config.ts
@@ -1,0 +1,5 @@
+import { Dimensions } from './constants';
+
+export const Config = {
+    defaultDimension: Dimensions.MD,
+};